clean up code structure
